+++ lib/openapi_first.rb
@@ -2,7 +2,7 @@
 
 require 'yaml'
 require 'json_refs'
-require_relative 'openapi_first/config'
+require_relative 'openapi_first/configuration'
 require_relative 'openapi_first/plugins'
 require_relative 'openapi_first/definition'
 require_relative 'openapi_first/version'
@@ -11,10 +11,10 @@ require_relative 'openapi_first/router'
 require_relative 'openapi_first/request_validation'
 require_relative 'openapi_first/response_validator'
 require_relative 'openapi_first/response_validation'
-require_relative 'openapi_first/error_responses/default'
-require_relative 'openapi_first/error_responses/json_api'
 
 module OpenapiFirst
+  extend Plugins
+
   # The OpenAPI operation for the current request
   OPERATION = 'openapi.operation'
 
@@ -36,12 +36,22 @@ module OpenapiFirst
   # The parsed request body
   REQUEST_BODY = 'openapi.parsed_request_body'
 
-  def self.load(spec_path, only: nil)
-    resolved = Dir.chdir(File.dirname(spec_path)) do
-      content = YAML.load_file(File.basename(spec_path))
-      JsonRefs.call(content, resolve_local_ref: true, resolve_file_ref: true)
+  class << self
+    def configuration
+      @configuration ||= Configuration.new
     end
-    resolved['paths'].filter!(&->(key, _) { only.call(key) }) if only
-    Definition.new(resolved, spec_path)
+
+    def configure
+      yield(configuration)
+    end
+
+    def load(spec_path, only: nil)
+      resolved = Dir.chdir(File.dirname(spec_path)) do
+        content = YAML.load_file(File.basename(spec_path))
+        JsonRefs.call(content, resolve_local_ref: true, resolve_file_ref: true)
+      end
+      resolved['paths'].filter!(&->(key, _) { only.call(key) }) if only
+      Definition.new(resolved, spec_path)
+    end
   end
 end