+++ lib/openapi_first/plugins/jsonapi.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require_relative 'jsonapi/error_response'
+
+module OpenapiFirst
+  module Plugins
+    module Jsonapi
+      OpenapiFirst.register(:jsonapi, self)
+    end
+  end
+end