+++ lib/openapi_first/request_validation.rb
@@ -28,6 +28,13 @@ module OpenapiFirst
       )
     end
 
+    def self.find_error_response(mod)
+      return OpenapiFirst.configuration.error_response if mod.nil?
+      return OpenapiFirst.plugin(mod)::ErrorResponse if mod.is_a?(Symbol)
+
+      mod
+    end
+
     # @param app The parent Rack application
     # @param options An optional Hash of configuration options to override defaults
     #   :error_response A Boolean indicating whether to raise an error if validation fails.
@@ -36,9 +43,9 @@ module OpenapiFirst
     #                   default: false (Config.default_options.request_validation_raise_error)
     def initialize(app, options = {})
       @app = app
-      @raise = options.fetch(:raise_error, Config.default_options.request_validation_raise_error)
+      @raise = options.fetch(:raise_error, OpenapiFirst.configuration.request_validation_raise_error)
       @error_response_class =
-        Plugins.find_error_response(options.fetch(:error_response, Config.default_options.error_response))
+        self.class.find_error_response(options[:error_response])
     end
 
     def call(env)