+++ spec/plugins/jsonapi.rb/error_response_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-RSpec.describe OpenapiFirst::ErrorResponses::JsonApi do
+RSpec.describe OpenapiFirst.plugin(:jsonapi)::ErrorResponse do
   describe '#render' do
     let(:env) { {} }